Have separate spec builder for each platform
