+++ pkg/cri/sbserver/container_create_windows.go
@@ -23,56 +23,12 @@ import (
 	runtime "k8s.io/cri-api/pkg/apis/runtime/v1"
 
 	"github.com/containerd/containerd/oci"
-	customopts "github.com/containerd/containerd/pkg/cri/opts"
 	"github.com/containerd/containerd/snapshots"
 )
 
 // No container mounts for windows.
 func (c *criService) containerMounts(sandboxID string, config *runtime.ContainerConfig) []*runtime.Mount {
 	return nil
-}
-
-func (c *criService) platformSpec(
-	id string,
-	sandboxID string,
-	config *runtime.ContainerConfig,
-	sandboxConfig *runtime.PodSandboxConfig,
-	imageConfig *imagespec.ImageConfig,
-	extraMounts []*runtime.Mount,
-) ([]oci.SpecOpts, error) {
-	specOpts := []oci.SpecOpts{}
-
-	specOpts = append(specOpts,
-		customopts.WithWindowsMounts(c.os, config, extraMounts),
-		customopts.WithDevices(config),
-	)
-
-	// Start with the image config user and override below if RunAsUsername is not "".
-	username := imageConfig.User
-
-	windowsConfig := config.GetWindows()
-	if windowsConfig != nil {
-		specOpts = append(specOpts, customopts.WithWindowsResources(windowsConfig.GetResources()))
-		securityCtx := windowsConfig.GetSecurityContext()
-		if securityCtx != nil {
-			runAsUser := securityCtx.GetRunAsUsername()
-			if runAsUser != "" {
-				username = runAsUser
-			}
-			cs := securityCtx.GetCredentialSpec()
-			if cs != "" {
-				specOpts = append(specOpts, customopts.WithWindowsCredentialSpec(cs))
-			}
-		}
-	}
-
-	// There really isn't a good Windows way to verify that the username is available in the
-	// image as early as here like there is for Linux. Later on in the stack hcsshim
-	// will handle the behavior of erroring out if the user isn't available in the image
-	// when trying to run the init process.
-	specOpts = append(specOpts, oci.WithUser(username))
-
-	return specOpts, nil
 }
 
 // No extra spec options needed for windows.
