+++ pkg/cri/sbserver/helpers_windows.go
@@ -166,3 +166,7 @@ func ensureRemoveAll(_ context.Context, dir string) error {
 func modifyProcessLabel(runtimeType string, spec *specs.Spec) error {
 	return nil
 }
+
+func isUnifiedCgroupsMode() bool {
+	return false
+}
